// ScoutRoute.Frontend/scoutroute-driver/src/models/teamDto.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * ScoutRoute.ApiService | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ListRouteDto } from "./listRouteDto";

export interface TeamDto {
  id?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  teamLead?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  trailerType?: number | null;
  members?: string[];
  routes?: ListRouteDto[];
}